update team photo via interaction
